Define IPool interface
