+++ src/index.ts
@@ -1,19 +1,19 @@
-import { DynamicClusterPool } from './pools/cluster/dynamic'
-import { FixedClusterPool } from './pools/cluster/fixed'
-import { DynamicThreadPool } from './pools/thread/dynamic'
-import { FixedThreadPool } from './pools/thread/fixed'
-import { ClusterWorker } from './worker/cluster-worker'
-import { ThreadWorker } from './worker/thread-worker'
-
+export { AbstractPool } from './pools/abstract-pool'
 export type { PoolOptions } from './pools/abstract-pool'
+export { DynamicClusterPool } from './pools/cluster/dynamic'
+export { FixedClusterPool } from './pools/cluster/fixed'
 export type { ClusterPoolOptions } from './pools/cluster/fixed'
+export type {
+  ErrorHandler,
+  ExitHandler,
+  IPool,
+  IWorker,
+  OnlineHandler
+} from './pools/pool'
+export { DynamicThreadPool } from './pools/thread/dynamic'
+export { FixedThreadPool } from './pools/thread/fixed'
 export type { ThreadWorkerWithMessageChannel } from './pools/thread/fixed'
+export { AbstractWorker } from './worker/abstract-worker'
+export { ClusterWorker } from './worker/cluster-worker'
+export { ThreadWorker } from './worker/thread-worker'
 export type { WorkerOptions } from './worker/worker-options'
-export {
-  FixedThreadPool,
-  FixedClusterPool,
-  DynamicClusterPool,
-  DynamicThreadPool,
-  ThreadWorker,
-  ClusterWorker
-}
